Areaofcircle.py
Here we are finding the area of circle.

+++ Areaofcircle.py
@@ -1,4 +1,4 @@
-radius = int(input("input the radius of the circle"))
+radius = float(input("input the radius of the circle"))
 area = 3.14 * radius * radius 
 print('the area of the circle is : ',area)
